requirements-dev.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
ruff>=0.2.0
mypy>=1.8.0
bandit>=1.7.6
pyyaml>=6.0.1
